Add user and date fields to posts feature
Implemented migration to add a user reference to the posts table for associating posts with specific users.
Added a date field to the posts table to track when posts are created.
Updated the events_controller and users_controller to incorporate changes.
Modified index.html.erb and profile.html.erb views to display relevant updates.